validate: Move application code to validate/application package
Same pattern as the clusters move. application.Command embeds
*validatecmd.Command, facade delegates to application.NewCommand().

The thin wrapper Command in validate/command.go and all shared test
helpers in validate_test.go are no longer needed — validate/ is now
just the facade with a single file.
