+++ pkg/validate/application/command_validated_test.go
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: The RamenDR authors
 // SPDX-License-Identifier: Apache-2.0
 
-package validate
+package application
 
 import (
 	"fmt"
